Fix robocop linter errors
